fix(interp): align jit profiling lifecycle
